quicksight-automation/requirements.txt
boto3==1.17.23
botocore==1.20.23
botostubs==0.14.1.17.20
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.3
